+++ package.json
@@ -17,7 +17,10 @@
     "react-icons": "^4.6.0"
   },
   "devDependencies": {
+    "autoprefixer": "^10.4.13",
     "eslint": "8.26.0",
-    "eslint-config-next": "13.0.0"
+    "eslint-config-next": "13.0.0",
+    "postcss": "^8.4.18",
+    "tailwindcss": "^3.2.1"
   }
 }
