+++ usersData/LOG
@@ -1,3 +1,3 @@
-2022/11/01-10:33:37.288 158c Recovering log #377
-2022/11/01-10:33:37.378 158c Delete type=0 #377
-2022/11/01-10:33:37.378 158c Delete type=3 #376
+2022/11/01-12:55:19.995 32d4 Recovering log #421
+2022/11/01-12:55:20.219 32d4 Delete type=0 #421
+2022/11/01-12:55:20.219 32d4 Delete type=3 #420
