+++ usersData/LOG.old
@@ -1,3 +1,3 @@
-2022/11/01-10:31:02.221 12c8 Recovering log #375
-2022/11/01-10:31:02.327 12c8 Delete type=0 #375
-2022/11/01-10:31:02.327 12c8 Delete type=3 #374
+2022/11/01-12:55:10.796 2178 Recovering log #419
+2022/11/01-12:55:10.896 2178 Delete type=0 #419
+2022/11/01-12:55:10.897 2178 Delete type=3 #418
